test: Collect compileModule tests into one file
